Move pass manager to own file
