+++ vllm/compilation/backends.py
@@ -6,7 +6,6 @@ from unittest.mock import patch
 
 import torch
 import torch.fx as fx
-from torch._inductor.codecache import BypassFxGraphCache
 
 import vllm.envs as envs
 from vllm.config import CompilationConfig
@@ -14,10 +13,7 @@ from vllm.logger import init_logger
 from vllm.utils import weak_ref_tensors
 
 from .counter import compilation_counter
-from .functionalization import FixFunctionalizationPass
-from .fusion import FusionPass
-from .inductor_pass import InductorPass, InductorPassType
-from .reshapes import RedundantReshapesPass
+from .pass_manager import PostGradPassManager
 
 logger = init_logger(__name__)
 
@@ -170,74 +166,6 @@ class PiecewiseCompileInterpreter(torch.fx.Interpreter):
             compilation_counter.num_piecewise_capturable_graphs_seen += 1
 
         return output
-
-
-class PostGradPassManager:
-    """
-    The pass manager for post-grad passes.
-    It handles configuration, adding custom passes, and running passes.
-    It also supports pickling, which is used by the Inductor code cache.
-    TODO in torch==2.6, use CustomGraphPass
-    (torch._inductor.custom_graph_pass.CustomGraphPass)
-
-    The order of the post-grad post-passes is:
-    1. passes (constructor parameter)
-    2. default passes (RedundantReshapesPass, FusionPass)
-    3. config["post_grad_custom_post_pass"] (if it exists)
-    4. fix_functionalization
-    This way, all passes operate on a functionalized graph.
-    """
-
-    def __init__(self):
-        self.passes: List[InductorPassType] = []
-
-    def __call__(self, graph: fx.Graph):
-        for pass_ in self.passes:
-            pass_(graph)
-
-        # always run fix_functionalization last
-        self.fix_functionalization(graph)
-
-    def configure(self, pass_config: CompilationConfig.PassConfig):
-        self.pass_config = pass_config
-        if pass_config.enable_reshape:
-            self.passes += [RedundantReshapesPass(pass_config)]
-
-        if pass_config.enable_fusion:
-            self.passes += [FusionPass.instance(pass_config)]
-
-        self.fix_functionalization = FixFunctionalizationPass(pass_config)
-
-    def add(self, pass_: InductorPassType):
-        self.passes.append(pass_)
-
-    def __getstate__(self):
-        """
-        Custom pickling for the pass manager, as some passes cannot be pickled.
-        Pickling occurs because the pass manager is set as the value of
-        `config["post_grad_custom_post_pass"]` in the Inductor config.
-
-        TODO in torch==2.6, use the `uuid` method in CustomGraphPass instead.
-        """
-        state = {"pass_config": self.pass_config}
-        passes_state = []
-        for pass_ in self.passes:
-            if isinstance(pass_, InductorPass):
-                passes_state.append(pass_.uuid())
-            else:
-                logger.warning("Custom pass does not inherit from "
-                               "InductorPass. Bypassing Inductor code cache.")
-                raise BypassFxGraphCache(
-                    "Custom pass should inherit InductorPass")
-        passes_state.append(self.fix_functionalization.uuid())
-        return state
-
-    def __setstate__(self, state):
-        """
-        Do not allow unpickling of the pass manager.
-        If this is needed in the future, it should properly pickle the passes.
-        """
-        raise ValueError("Cannot unpickle PostGradPassManager")
 
 
 class VllmBackend:
